feat: use Bearer token if OPA started with it #7340
