+++ jans-auth-server/common/src/main/java/io/jans/as/common/service/common/ConfigurationService.java
@@ -10,6 +10,7 @@ import io.jans.as.model.config.StaticConfiguration;
 import io.jans.as.persistence.model.configuration.GluuConfiguration;
 import io.jans.model.SmtpConfiguration;
 import io.jans.orm.PersistenceEntryManager;
+import io.jans.service.EncryptionService;
 import io.jans.util.StringHelper;
 import io.jans.util.security.StringEncrypter.EncryptionException;
 import org.slf4j.Logger;
