+++ jans-keycloak-link/server/src/main/java/io/jans/keycloak/link/server/service/AppInitializer.java
@@ -18,11 +18,11 @@ import com.google.common.collect.Lists;
 import io.jans.exception.ConfigurationException;
 import io.jans.keycloak.link.service.config.ApplicationFactory;
 import io.jans.keycloak.link.service.config.ConfigurationFactory;
-import io.jans.link.service.EncryptionService;
 import io.jans.model.custom.script.CustomScriptType;
 import io.jans.orm.PersistenceEntryManager;
 import io.jans.orm.model.PersistenceConfiguration;
 import io.jans.orm.util.properties.FileConfiguration;
+import io.jans.service.EncryptionService;
 import io.jans.service.PythonService;
 import io.jans.service.cdi.event.ApplicationInitialized;
 import io.jans.service.cdi.event.ApplicationInitializedEvent;
