+++ jans-keycloak-link/server/src/main/java/io/jans/keycloak/link/server/service/KeycloakService.java
@@ -3,8 +3,8 @@ package io.jans.keycloak.link.server.service;
 import io.jans.keycloak.link.model.config.CacheRefreshConfiguration;
 import io.jans.keycloak.link.model.config.KeycloakConfiguration;
 import io.jans.keycloak.link.service.config.ConfigurationFactory;
-import io.jans.link.service.EncryptionService;
 import io.jans.link.util.PropertyUtil;
+import io.jans.service.EncryptionService;
 import io.jans.util.security.StringEncrypter;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.enterprise.inject.Produces;
