+++ jans-lock/service/src/main/java/io/jans/lock/model/config/OpaConfiguration.java
@@ -20,12 +20,22 @@ public class OpaConfiguration implements Configuration {
 
 	private String baseUrl;
 
+	private String accessToken;
+
 	public String getBaseUrl() {
 		return baseUrl;
 	}
 
 	public void setBaseUrl(String baseUrl) {
 		this.baseUrl = baseUrl;
+	}
+
+	public String getAccessToken() {
+		return accessToken;
+	}
+
+	public void setAccessToken(String accessToken) {
+		this.accessToken = accessToken;
 	}
 
 }
